requirements.txt
keras==2.8.0
matplotlib==3.5.1
mpl_toolkits.clifford==0.0.3
numpy==1.21.6
pandas==1.4.3
Pillow==9.4.0
scikit_learn==1.2.1
seaborn==0.11.2
spacy==3.3.1
streamlit==1.13.0
